GoogleMapsUrlConverter: move to pkg converters
